fix: refactor ActivitiMessagingDestinationAutoConfiguration
