+++ activiti-cloud-service-common/activiti-cloud-service-messaging-config/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-    org.activiti.cloud.common.messaging.config.ActivitiCloudMessagingAutoConfiguration
+    org.activiti.cloud.common.messaging.config.ActivitiCloudMessagingAutoConfiguration,\
+    org.activiti.cloud.common.messaging.config.ActivitiMessagingDestinationsAutoConfiguration
