+++ activiti-cloud-service-common/activiti-cloud-service-messaging-config/src/main/resources/activiti-cloud-messaging.properties
@@ -5,3 +5,10 @@ activiti.cloud.messaging.instance-index=${ACT_MESSAGING_INSTANCE_INDEX:0}
 activiti.cloud.messaging.destination-separator=${ACT_MESSAGING_DEST_SEPARATOR:_}
 activiti.cloud.messaging.destination-prefix=${ACT_MESSAGING_DEST_PREFIX:}
 activiti.cloud.messaging.destination-override-enabled=${ACT_MESSAGING_DEST_OVERRIDE_ENABLED:false}
+
+# Default destination scope mappings for cloud stream bindings with matching channel destination
+activiti.cloud.messaging.destinations.commandConsumer.scope=${activiti.cloud.application.name}
+activiti.cloud.messaging.destinations.engineEvents.scope=
+activiti.cloud.messaging.destinations.asyncExecutorJobs.scope=${activiti.cloud.application.name}
+activiti.cloud.messaging.destinations.messageEvents.scope=${activiti.cloud.application.name}
+activiti.cloud.messaging.destinations.commandResults.scope=${activiti.cloud.application.name}
