+++ activiti-cloud-service-common/activiti-cloud-service-messaging-starter/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,2 @@
 org.springframework.boot.env.EnvironmentPostProcessor=\
   org.activiti.cloud.common.messaging.config.ActivitiMessagingEnvironmentPostProcessor
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-    org.activiti.cloud.common.messaging.config.ActivitiMessagingDestinationsAutoConfiguration
-
